Refactor and simplify (#59981)
