--- src/EditorFeatures/Core/InfoBar/EditorInfoBarService.cs
@@ -1,27 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-#nullable disable
-
-using System;
-using System.Composition;
-using Microsoft.CodeAnalysis.Extensions;
-using Microsoft.CodeAnalysis.Host.Mef;
-using Microsoft.CodeAnalysis.Internal.Log;
-
-namespace Microsoft.CodeAnalysis.Editor.Implementation.Workspaces
-{
-    [ExportWorkspaceService(typeof(IInfoBarService)), Shared]
-    internal class EditorInfoBarService : IInfoBarService
-    {
-        [ImportingConstructor]
-        [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-        public EditorInfoBarService()
-        {
-        }
-
-        public void ShowInfoBar(string message, params InfoBarUI[] items)
-            => Logger.Log(FunctionId.Extension_InfoBar, message);
-    }
-}
