--- src/EditorFeatures/Core/Workspaces/EditorErrorReportingServiceFactory.cs
@@ -1,29 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-#nullable disable
-
-using System;
-using System.Composition;
-using Microsoft.CodeAnalysis.Extensions;
-using Microsoft.CodeAnalysis.Host;
-using Microsoft.CodeAnalysis.Host.Mef;
-
-namespace Microsoft.CodeAnalysis.Editor.Implementation.Workspaces
-{
-    [ExportWorkspaceServiceFactory(typeof(IErrorReportingService), ServiceLayer.Editor), Shared]
-    internal class EditorErrorReportingServiceFactory : IWorkspaceServiceFactory
-    {
-        private readonly IErrorReportingService _singleton = new EditorErrorReportingService();
-
-        [ImportingConstructor]
-        [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-        public EditorErrorReportingServiceFactory()
-        {
-        }
-
-        public IWorkspaceService CreateService(HostWorkspaceServices workspaceServices)
-            => _singleton;
-    }
-}
