+++ src/VisualStudio/Core/Def/ErrorReporting/DetailedErrorInfoDialog.xaml
@@ -1,8 +1,8 @@
-﻿<ui:DialogWindow x:Class="Microsoft.VisualStudio.LanguageServices.Implementation.DetailedErrorInfoDialog"
+﻿<ui:DialogWindow x:Class="Microsoft.CodeAnalysis.ErrorReporting.DetailedErrorInfoDialog"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
-             xmlns:local="clr-namespace:Microsoft.VisualStudio.LanguageServices.Implementation"
+             xmlns:local="clr-namespace:Microsoft.CodeAnalysis.ErrorReporting"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:ui="clr-namespace:Microsoft.VisualStudio.PlatformUI;assembly=Microsoft.VisualStudio.Shell.15.0"
              xmlns:vs="clr-namespace:Microsoft.VisualStudio.Shell;assembly=Microsoft.VisualStudio.Shell.15.0"
