+++ src/VisualStudio/Core/Def/ErrorReporting/VisualStudioErrorReportingService.ExceptionFormatting.cs
@@ -10,9 +10,10 @@ using System.Linq;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Text;
+using Microsoft.VisualStudio.LanguageServices;
 using StreamJsonRpc;
 
-namespace Microsoft.VisualStudio.LanguageServices.Implementation
+namespace Microsoft.CodeAnalysis.ErrorReporting
 {
     internal partial class VisualStudioErrorReportingService
     {
